+++ service/config/runtimeconfig.go
@@ -1,6 +1,8 @@
 package config
 
 import (
+	"time"
+
 	"github.com/streamingfast/dstore"
 
 	"github.com/streamingfast/substreams/orchestrator/work"
@@ -11,9 +13,11 @@ import (
 type RuntimeConfig struct {
 	StateBundleSize uint64
 
-	MaxWasmFuel                uint64 // if not 0, enable fuel consumption monitoring to stop runaway wasm module processing forever
-	MaxJobsAhead               uint64 // limit execution of depencency jobs so they don't go too far ahead of the modules that depend on them (ex: module X is 2 million blocks ahead of module Y that depends on it, we don't want to schedule more module X jobs until Y caught up a little bit)
-	DefaultParallelSubrequests uint64 // how many sub-jobs to launch for a given user
+	MaxWasmFuel        uint64        // if not 0, enable fuel consumption monitoring to stop runaway wasm module processing forever
+	MaxJobsAhead       uint64        // limit execution of depencency jobs so they don't go too far ahead of the modules that depend on them (ex: module X is 2 million blocks ahead of module Y that depends on it, we don't want to schedule more module X jobs until Y caught up a little bit)
+	InitSubrequests    int           // how many sub-jobs to start exactly when the tier1 request comes in
+	SubrequestsRampup  time.Duration // over this amount of time, parallel jobs count will go from 'InitParallelSubrequests' to the 'ParallelSubrequests' (default or overriden per request)
+	DefaultSubrequests int           // how many sub-jobs to launch for a given user
 	// derives substores `states/`, for `store` modules snapshots (full and partial)
 	// and `outputs/` for execution output of both `map` and `store` module kinds
 	BaseObjectStore dstore.Store
@@ -26,6 +30,8 @@ type RuntimeConfig struct {
 func NewRuntimeConfig(
 	stateBundleSize uint64,
 	parallelSubrequests uint64,
+	initParallelSubrequests uint64,
+	parallelRampupPeriod time.Duration,
 	maxJobsAhead uint64,
 	maxWasmFuel uint64,
 	baseObjectStore dstore.Store,
@@ -33,13 +39,15 @@ func NewRuntimeConfig(
 	workerFactory work.WorkerFactory,
 ) RuntimeConfig {
 	return RuntimeConfig{
-		StateBundleSize:            stateBundleSize,
-		DefaultParallelSubrequests: parallelSubrequests,
-		MaxJobsAhead:               maxJobsAhead,
-		MaxWasmFuel:                maxWasmFuel,
-		BaseObjectStore:            baseObjectStore,
-		DefaultCacheTag:            defaultCacheTag,
-		WorkerFactory:              workerFactory,
+		StateBundleSize:    stateBundleSize,
+		DefaultSubrequests: int(parallelSubrequests),
+		InitSubrequests:    int(initParallelSubrequests),
+		SubrequestsRampup:  parallelRampupPeriod,
+		MaxJobsAhead:       maxJobsAhead,
+		MaxWasmFuel:        maxWasmFuel,
+		BaseObjectStore:    baseObjectStore,
+		DefaultCacheTag:    defaultCacheTag,
+		WorkerFactory:      workerFactory,
 		// overridden by Tier Options
 		ModuleExecutionTracing: false,
 	}